fix(codegen): Map math.isqrt to inline sqrt implementation (Refs DEPYLER-0771)
Python's math.isqrt() returns integer square root. Since Rust's f64 doesn't
have an isqrt method, we generate inline code: (n as f64).sqrt().floor() as i32

Changes:
- Add isqrt to module_mapper.rs and module_mapper_phf.rs item maps
- Add inline conversion in expr_gen.rs for both method and imported calls
- Skip invalid std::f64::isqrt import generation in rust_gen.rs

Also includes accumulated example regenerations and test updates.

+++ examples/test_docs_simple.rs
@@ -16,7 +16,7 @@ impl DataProcessor {
         self.data.push(value);
     }
     pub fn get_count(&self) -> i32 {
-        return self.data.len();
+        return self.data.len() as i32;
     }
     pub fn create_default() -> DataProcessor {
         return DataProcessor::new("default".to_string());
